Split out profile example into separate files, one for each page
